requirements.txt
# Framework agente
agno[mcp,postgres]>=2.5
anthropic>=0.40

# Memoria e Knowledge
supermemory>=1.0

# Telegram
aiogram>=3.13

# STT / TTS
deepgram-sdk>=3.8
elevenlabs>=1.15

# Database
asyncpg>=0.30
psycopg2-binary>=2.9

# Scheduler
APScheduler>=3.10

# Config
pydantic-settings>=2.6
python-dotenv>=1.0

# Utilities
httpx>=0.27

# Testing
pytest>=8.0
pytest-asyncio>=0.24
